Update README.md
This is a small project for the WebeElevate - Web and Mobile App development course.

The goal is to make mobile app mockup, which can be viewed online though a browser.  It should present a number of screens 
that are clickable, giving a the user the impression of a functioning app, where possible.